requirements.txt
torch>=2.0.0
numpy>=1.24
pretty_midi>=0.2.10
music21>=9.1
pyyaml>=6.0
tqdm>=4.65
pandas>=2.0
